Updated Projectile Handler
Updated enemy projectile handler to accept attack data from the state and to update the projectile handler status based on whether it is in progress or complete.

Added game event SO for winning and losing the game and tied them to their respective player/enemy take damage method. Triggered when health reaches 0 for either.