Fixed enemy repeat attack bug
There was an issue with the changing of the player controller enum card selection . It was occurring in the Decision SO for the enemy sates. Might have been a value/reference issue. Changed the var to refer to the class and it seems to have fixed the issue with the enum variable not changing.

Added additional player controller enum for playerStatus to track whether player is isIdle or inProgress