Spell Projectile Update
Updated the movement to not only start moving when the vector forward is lined up with the enemy. Moves straight towards the enemy position. Added a fixed Y amount to the target position so that it more consistently hits the enemy target.

Updated the enemy projectile handler to try and address an issue with console errors dealing with index for multi enemy projectile shots